move tools into a separate magefiles module
