+++ magefiles/buf.gen.yaml
@@ -4,26 +4,26 @@ version: "v1"
 plugins:
   - name: 'go'
     path: ["go", "run", "google.golang.org/protobuf/cmd/protoc-gen-go"]
-    out: 'proto'
+    out: '../proto'
     opt: 'paths=source_relative'
   - name: 'go-grpc'
-    out: 'proto'
+    out: '../proto'
     path: ["go", "run", "google.golang.org/grpc/cmd/protoc-gen-go-grpc"]
     opt: 'paths=source_relative'
   - name: 'go-vtproto'
     path: ["go", "run", "github.com/planetscale/vtprotobuf/cmd/protoc-gen-go-vtproto"]
-    out: 'proto'
+    out: '../proto'
     opt: 'paths=source_relative,features=marshal+unmarshal+size+clone'
   - name: 'validate'
     path: ["go", "run", "github.com/envoyproxy/protoc-gen-validate"]
-    out: 'proto'
+    out: '../proto'
     opt: 'paths=source_relative,lang=go'
   - name: "grpc-gateway"
     path: ["go", "run", "github.com/grpc-ecosystem/grpc-gateway/v2/protoc-gen-grpc-gateway"]
-    out: "proto"
+    out: "../proto"
     opt: "paths=source_relative"
   - name: "openapiv2"
-    out: "proto"
+    out: "../proto"
     path: ["go", "run", "github.com/grpc-ecosystem/grpc-gateway/v2/protoc-gen-openapiv2"]
     strategy: "all"
     opt: "allow_merge=true"
