+++ magefiles/deps.go
@@ -2,15 +2,20 @@
 
 package main
 
-import "github.com/magefile/mage/mg"
+import (
+	"fmt"
 
-var goModules = []string{"."}
+	"github.com/magefile/mage/mg"
+)
+
+var goModules = []string{".", "magefiles"}
 
 type Deps mg.Namespace
 
 // Tidy go mod tidy all go modules
 func (Deps) Tidy() error {
 	for _, mod := range goModules {
+		fmt.Println("tidying", mod)
 		if err := runDirV(mod, "go", "mod", "tidy"); err != nil {
 			return err
 		}
@@ -22,6 +27,7 @@ func (Deps) Tidy() error {
 // Update go get -u all go dependencies
 func (Deps) Update() error {
 	for _, mod := range goModules {
+		fmt.Println("updating", mod)
 		if err := runDirV(mod, "go", "get", "-u", "-t", "-tags", "tools", "./..."); err != nil {
 			return err
 		}
